Use get_success in tests handling `@cached` funcs
As of [1], successResultOf is typed. It expects a Deferred. We give it
the return type `T` of asynchronous function which is decorated by @cached.
I.e. `T` is the type of `@cached(func)(arg, ...)`. AFAICS `T` is a
`DeferredCacheDescriptor` which sounds like something that ought to
produce `Deferred`s... but again, it's mysterious murky magic.

Use `get_success` instead, which seems to be the done thing
elsewhere. (It internally uses `successResultOf` anyway.)

[1]: https://github.com/twisted/twisted/pull/1736/commits/3f8c49816bc34f74167ea3e0579f0d5dde25aeec