Reconcile state when processing incoming events during faster joins
In order to handle incoming events over federation during a faster join, we
need to relax the auth rules. Specifically, we need to accept that we may not
have the sender's membership in our view of the room state. However, it should
be in the `auth_events`, and state-resolving the auth events against our view
should give the right thing.