Use `AsyncLruCache` for the get events cache
This doesn't change function of the cache at all but prepares for the
possibility of external backed, or combination external/local backed,
event cache.

This does separate the cache invalidation into two options: the local
in memory cache and any asynchronous cache (which is currently the
same thing). This means that the worker invalidating objects on changes
must call the async method and the local method is reserved for
invalidations received over replication.