Optionally disable state groups caching
The theory behind this is that the overhead and high churn rate of the
cache result in greater CPU and memory wastage than is offloaded from
the database.

Note this still caches groups on insert which means multiple room inserts
in a short period (ie during room creation) will still benefit from the
cache.