Fixup the dockerfiles to the latest poetry updates
This was a total mess because for some reason setuptools isn't available
at or during the requirements install phase. For the non-pyston image
this was just a case of installing the shared secret authenticator before
the rest of the packages. In the pyston image I have just excluded the
pinned setuptools package for now.