Add `executescript` to DB engines
So we can feed in a big SQL file to sql and sqlite without having to
worry about parsing it. Motivation: our handwritten sort=of-parser
doesn't like the CREATE TRIGGER statements (or at least the quoting it
uses) in the upcoming schema dump.

Irritatingly the `cursor` that the DB engines get isn't a raw cursor but
a `LoggingTransaction`. That is, the annotations I wrote on the DB
engine are wrong. Maybe we can make `LoggingTransaction` generic over a
cursor type in the future.