feat: support subdirs for appservice config files
The current Docker container only supports when registration files are placed in the /data/appservices directory. Subdirectories are not currently supported. In the case of a user running Synapse on Kubernetes, the cleanest way to install appservice configuration files would be to use an individual configmap for each different bridge, resulting in multiple subdirectories for each appservice. Supporting placing registration files in subdirectories makes this deployment usecase easier.

Signed-off-by: Danny Seymour <danny@seymour.family>