Ensure device inbox stream change caches are correctly updated for new IDs
This ensures that every time a new device inbox stream ID is generated
both stream change caches that utilise it are notified of that position,
even if there are no relevant changes at that point.

This ensures that the max position of these caches is always updated to
match the current token, which then makes it possible to wait for changes
to arrive over replication.