Update tests using 'update_user_directory'.
By setting 'update_user_directory_from_worker' to 'None', it disables the functionality(and I added that comment where the default config is written out). If you need to re-enable this functionality on master, either .pop() that value or alternatively set it to master which appears to work as well.