Setup a prewarmed docker build for Complement that will run in parallel with the pre-linting-done steps.
This will build the complement-synapse image(which usually takes between 3 and 5 minutes) at an earlier stage instead of adding that time to the Complement test run. Instead of caching this in Docker Hub or ghcr.io, upload it as an artifact that will be subsequently downloaded during the Complement test run. Additional detail: do the docker building on a tmpfs to reduce latency.