Remove job~= condition and aggregate with max()
The job~= condition is not portable: it depends on how we do things on EMS and matrix.org

Other people don't set up their job names the same way we do.

Therefore we can't rely on job~= and cut it out.

But doing that means that you end up with one trace per worker.

To avoid this being a problem, I propose we aggregate them with max() so we take the number from the

correct worker that is reporting MAU.