Check for stream position at or ahead when flagging entity as changed
This has to be here to account for applying changes on workers that
acually increment the stream ID - they have to increment the token before
they know about it to flag the change.

This does leave open a potential hole in the check for replicated changes
which should be addressed in the future.