Add check to avoid starting duplicate partial state syncs
When we are kicked from a room while syncing partial state, the sync
can fail and leave us in a state where the homeserver is not joined to
the room, yet has the partial state flag set on the room.

When we next join the room (implemented by Mat in a separate PR), we
would like to restart the sync. To do so without starting duplicate
syncs, we add a set which tracks which partial state syncs are in
progress.

Signed-off-by: Sean Quah <seanq@matrix.org>