Change ID generator to return position of last write
This is already true when asking for stream positions of other
instances, but for our own instance we have fudged it. Changing this
should be fine (it was just an optimisation), and I don't think it
should have much impact in practice at all.

The reason to do this is so that when tell remotes what our current
position is we only include *our* writes, rather than writes of other
instances. This reduces delays when the remote instance is waiting for
stream positions to update. In practice, this is probably only a problem
for tests, though we may as well do it for all of them.