Refactor `get_partial_state_servers_at_join`
Return `None` when the given room is no longer partial stated, to
explicitly indicate when the room has partial state. Otherwise it's not
clear whether an empty list means that the room has full state, or the
room is partial stated, but the server we joined off told us that there
are no servers in the room.

Signed-off-by: Sean Quah <seanq@matrix.org>